+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix broken app config API and CHANGELOG.md [#401](https://github.com/rokwire/core-building-block/issues/401)
 
 ### Changed
+- Update the filter capability for the get login sessions Admin API [#394](https://github.com/rokwire/core-building-block/issues/394)
 - Limit the returned items for get application accounts admin API [#375](https://github.com/rokwire/core-building-block/issues/375)
 - Return ordered list of account auth types on link account auth type [#376](https://github.com/rokwire/core-building-block/issues/376)